requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.3
drf-yasg==1.21.7
requests==2.31.0
celery==5.3.6
django-celery-results==2.5.1
django-debug-toolbar==4.3.0